Revert device.* changes in favor of `deCONZ::ApsController::instance()` usage
